Added missing assume in exercises/ex1a-safe-read-write.fst
In general exercises don't need to type-check, only solutions do.
But in this case we refer to the file from other points, so better
to have it compile.
The better long term solution is to only extract what's needed
each time, but for that we need proper support for literate programming.