wip on inversion lemmas and some little things
I replaced VDistinctC with 3 rules, because the conclusion
of VDistinctC needs to know the type of the constants.

Near the end of the file, I think Fstar gives an inconsistent type
for at some point in the code in comment. See if you can reproduce
this example. I do not have time to minimize it right now …