fixed the pattern-extraction issues exposed in test cases (naryTree).
the non-type pattern variables also had to be skipped just as the type pattern vars were skipped.
() is returned when the F* argument of a constructor is missing in ML.
This is done by putting () in the context when processing a skipped term pattern var

contrary to an earlier commit message, the type schemes of indexed inductive types are being computed correctly.
naryTree does not have any indices.