More tweaks on the dependency analysis.
- Files found in one of the include directories (excluding the current working
  directory) are printed out with absolute paths.
- Files found in the current directory are printed as relative paths.
- Files passed on the command-line are kept "as-is".

This provides the nice property that any path output by the --dep tool can be
passed to another tool, as long as you don't move to another directory.

This is also what I believe a Makefile consumer would want (so that they don't
have to re-implement the logic of "let's search for the file in the include
path").

I can easily add an option to change this behavior if needed.