Small change of semantics for auto-deps and verify-module.
The behavior is now as follows, in auto-deps.
- If the user didn't use --verify_module, we take the filenames passed on the
command-line to be all the modules the user wants to see verified.
- If the user used --verify_module, then we don't do anything smart and use the
user-provided list of modules to be verified.

The behavior remains unchanged in explicit-deps.
- If the user didn't use --verify_module, we verify every module passed on the
command-line.
- If the user used --verify_module, we verify those modules.

The advantage of the default behavior of auto-deps is only the modules you care
about are verified; their dependencies are not.