typechecker/prims: drop wp_as_type in favor of wp_stronger
We're really checking for wether one WP is stronger than the other, so
do that instead of going via wp_binop and wp_as_type. (Also note that
wp_binop of ==> will not generate a real WP, meaning monotonic and etc,
so having that intermediate step is not too nice).