dep: allow 'let open' on module abbrevs
At some point we would like to treat 'M t' as 'let open M in t',
but the latter requires M to be a fully qualified module name,
not a module abbrev, whereas examples do use 'M t' where M is a
module abbrev.

This commit alone has not been tested thoroughly, in the sense that
it may be the case that it allows more than module abbrevs.
It remains to test whether this commit (unexpectedly) makes the
following example work:

module Test
open A
let open B in ()

module A.B