Regenerate hints due to failed build 2578
At this point, Crypto.AEAD.Lemmas succeeds, but Ex07c still fails.
In fact, Ex07c succeeds if hints are disabled altogether, but even
if hints are produced when disabled (yielding the .hints file
as reflected in this commit), rechecking Ex07c with them
ultimately fails, for an unknown reason so far.