Move tosugar to syntax directory so it can be used in print.
Make a copy of constants and functions that are defined in syntax but
are used by parser into a seperate Parser.Const file so that we don't
need dependent on syntax project to build parser project. This allows us
to more resugar into Syntax directory so that it can be used by
syntax.print. This is temparory workaround. Eventually, we should move
the const into somewhere so we don't need two copies.