Generalize implementation of #show-match
The command now takes a type, not just an identifier.  That type can be a
abstraction, too (this is needed to allow F* to parse and typecheck types with
holes). Some examples:

>>> #show-match list
list
[]
(${hd<<<a|||(fun a -> a)>>>} :: ${tl<<<(list a)|||(fun a -> (list a))>>>})]

>>> #show-match list (int * int)
list
[]
(${hd<<<(tuple2 int int)|||(tuple2 int int)>>>} :: ${tl<<<(list (tuple2 int int))|||(list (tuple2 int int))>>>})]

>>> #show-match (fun a -> (a * int) * int)
tuple2
(${_1<<<(tuple2 a int)|||(fun a -> (tuple2 a int))>>>}, ${_2<<<int|||int>>>})]