Add .ignore files for search tools
ag and ripgrep both search `.ignore` files and do not search matched
files. These ignores files tell ag and ripgrep to ignore *.fst.hints
files and the source for reveal.js in the POPl 2015 slides, which need
to be checked in (so they aren't in .gitignore) but also shouldn't be
searched by default.