typechecker: refresh z3 for each goal
Definitely NOT OPTIMAL, but needed right now since we want to change
proof namespaces between goals.

A better way of doing this would be to check for which goals have the
same namespace, or extensions, and group them appropriately. But maybe a
more dynamic way of namespace handling would be better altogether.