tactics/normalizer: prevent unfolding of some logical connectors
Currently, normalizing turns many formulas unprovable. The following
failing assertions are a witness of that:

  let _ = assert_norm (True /\ True)

  let _ = assert (c_and True True)

  let _ = assert (c_and c_True c_True)

  val l1 : (a : Type) -> Lemma (a ==> squash a)
  let l1 a = assert_norm (a ==> squash a)

  val l2 : (a : Type) -> Lemma (squash a ==> a)
  let l2 a = assert_norm (squash a ==> a)

A proper fix for this is not trivial. We'll work around in tactics, for
now, by preventing the unfolding of the operators that are implemented
as squashes. Otherwise, runnning "simpl" or some other kind of
normalization tactic would render the goal unsolvable.