encoding/syntax: optimize unfolded logical connectives
We have p /\ q =def= squashs (c_and p q). However, the first one is
optimized by the encoding to an SMT2 `and`, and the second is not.

Allow for the encoding to detect and optimize such unfolding of logical
connectives by making destruct_typ_as_formula check for them.