Fix validator definition and correctness
A validator is a special case of a parser (one that returns a unit), so
re-use the definition. This also inherits the and_then combinator.

There was a bug in the correctness specification that actually made
`validate_many` incorrect: the validator must also consume the same
number of bytes as the parser, otherwise the composition of validators
doesn't work for the composition of parsers as their positions in the
input buffer might become desynchronized.