extraction: eta-expand coertions
If we have
	let x : 'a. 'a -> 'a = Obj.magic (...)
OCaml will complain because of the value restriction. A solution is to
eta-expand into `fun x -> Obj.magic (..) x`, which permits the
generalization.

A downside: this can introduce a warning:
	File "A.fst", line 5, characters 129-132:
	Warning 20: this argument will not be used by the function.
which is bogus. And can actually be avoided by not left-parenthesizing
applications (`(a b) c` vs `a b c`)

Somewhat related ref: https://caml.inria.fr/mantis/view.php?id=6660