Relax is_active_union_field; add an operation setting the tag of a C union
This tag does not actually exist at runtime, but in Arraystruct's model of the
heap, it does exist. Therefore, this operation is stateful, but should be
compiled to a no-op by Kremlin.

This is then used by the TaggedUnion library to enable a pattern where the user
can change the tag, and initialize the corresponding value in a separate steps.

Because of the new "is_active_union_field" relaxed definition, the user _does
not_ have to re-establish some invariant after doing so. The TaggedUnion is
"valid" at all instants, since we can now adjust the tag of a C union without
actually writing data in its field.