Revert "extraction: ocaml: fix annotation for top level lets" (partially)
(This reverts commit 339782a7975160fb149d19bc6dbcc9bcd1a931b3.)

The printers for ocaml 4.02.X and 4.04.X work OK just before, so there
was no real need for this fix. Even more, this breaks some programs for
ocaml 4.02.3 since, as explained in that commit, it caused declarations
to be parenthesized, like the following from miTLS:

	let (rkeys :
	  (Prims.unit,kset,(Prims.unit,Prims.unit) mon_pkey)
	    FStar_Monotonic_RRef.m_rref)
	  = FStar_Monotonic_RRef.m_alloc keyRegion []

curiously, those parentheses cause the declaration to fail with
"The type of this expression .. contains type variables that cannot be generalized",
which seems wrong, as there shouldn't be any difference introduced by
the ().

Further, OCaml 4.05.0 works just fine with both, so I'm assuming this is
a quirk in old OCaml versions that was fixed.

Hence, don't do anything for the known-good ocaml versions, and simply
fix the one for OCaml 4.05.0.