tactics: (1/2) Improve our bootstrapping story for native tactics
The compiler no longer depends on a file extracted from ulib.  Instead, we
isolated the one definition that the compiler needed (the __result type of the
tactic monad) an duplicated it in the compiler.

Ideally we'd have wanted it to live only in the compiler (then ulib could have
depended on that file), but that caused loadpath issues: every client of
FStar.Tactics had to add an extra folder to their loadpath to find
FStar.Tactics.Result.fs in the compiler's sources (we couldn't simply depend on
an interface in ulib because the DM4F implementation needs to see the actual
definition.

Co-Authored-By: @victor-dumitrescu