Do not check for snapshot fixpoint
We developers must be disciplined enough to ensure that we
shall have reached a fixpoint with OCaml 4.02.3, currently the
only version of OCaml for Windows that we tested so far.

However, we know that this check will fail with other versions of
OCaml, which we anyway need to use for anything other than
Windows.

We will restore this check once we switch to the latest OCaml on
Windows.