Consider the following code:
let rec f: a:Type u#x -> l:list u#x a -> list u#x a =
  fun a l -> []

Before this commit the code above did not verify because a universe substitution was not correctly applied to the implicit type argument to []. In master, this code works fine, because there a compress pass runs on the term that resolves the unification variable in the function body. That pass is not there in this branch, and so, the subsitution code skipped the unification variable. This commit has a bandaid solution to it.

***NO_CI***