Consider the following code:
let a n :nat = 1
let f (ls:list nat) :nat =
  let rec aux (xs:list nat) :nat = a 0
  in
  0

a has an implicit type parameter that is inferred to be ((fun (aux:t) -> int) aux) in the body of aux (there are other arguments to this lambda, but they are irrelevant for this commit).

In lax mode, the type t is simply (list nat -> Tot nat), because we don't add precedes guard in the lax mode.

In 2-phase, that creates a problem in the second phase, because the type of the aux argument there is guarded with precedes, whereas the lambda expects aux type to be without precedes.

This commit guards the type of recursive letbindings even in lax mode.

***NO_CI***