ide: (3/4) Get rid of most uses of commit_mark
commit_mark was a strange mix of a weird 'pop' (removing the second element on
the stack, the one introduced by calling 'mark' before 'push') and sometimes
what might have at some point been an actual commit, but is now just a pop (this
happens in FStar.SMTEncoding.Z3 and FStar.SMTEncoding.Encode).

Now that we got rid of 'mark' and 'reset_mark', we can also get rid of
most 'commit_mark's (the ones that only ever intended to pop).

'varops.commit_mark' in SMTEncoding.Encode is a special case: the code probably
intended to merge a new and an old environment, but it was wrong: due to the two
pushes, it always ended up merging an empty environment (pushed by 'mark') with
the non-empty topmost one before removing the empty one. This is the same as
popping the second environment directly, which we don't need to do anymore since
we got rid of 'mark'.

A similar problem exists in 'SMTEncoding.Z3': see the next commit.