ide: Clean up the backtracking implementation
The new implementation uses a global reference to the current repl stack; it was
implicit before (it pretended to be pure, but all operations on it had to be
paired with calls into TcEnv and DsEnv that touched global state).

The stack now contains the full repl state, ensuring that completion tables are
properly reverted.

tc_deps and update_deps do not create entries on that stack — instead, they have
their own little stack.  That's mostly because I have no idea what this code
does, and it does not conform to the usual push/pop abstraction.