normalizer: simplify: optimize wps on (fun _ -> True) postconditions
This is a common case when check satisfiability. Optimize:

 (forall x y. post x y) ==> E[post]
 ~> E[(fun x y -> True)]

This sound because of extensionality. Note that monotonicity is not
needed, otherwise we cannot do this since there's no requirement on E to
be a WP or anything.

This causes some variables to not appear,

 forall post. (forall x y. post x y) ==> E[post]
  ~> forall post. E [(fun _ _ -> True)]

possibly another patch incoming to fix, but needs to check for non-empty
domains or we're gonna prove False again, c.f. e2d2d619cb2b3916f67bcff8eb7bc21b6a8b9cb7