ide: Rework and document the way the IDE handles stacks
This also provides a better fix than 9cc7b43a for the issue at
https://github.com/FStarLang/fstar-mode.el/issues/82.  The previous fix did
this:

1. Push the current environment and get a copy
2. Run `task` with `st` containing the *old* environment (!)
3. Pop using the old environment (OK because `pop` only uses `env.solver`)
4. Store the environment copy from (1) in `st`, and return that

Concretely this worked — but that's not how the API was intended to be used.
Instead, the new solution pushes to get a copy of the environment, runs
computations in *that* fresh environment, and pops back to the original,
unmodified environment.

The added assertion only passes because of the fix in the previous commit.