tc: rel: make Tm_match unification fall back to SMT
Important: when trying to unify matches pointwise, we forbid SMT

Otherwise, consider

	match x with | 0 -> 42 | _ -> 42
	EQ
	match y with | 0 -> 42 | _ -> 42

they would not unify since we'd get a possibly unprovable (x = y) guard,
but the SMT can easily prove the top-level matches equivalent.

So, forbid SMT for the subproblems, and so if any guard is needed
internally we just generate a guard for the whole matches.