Bump version number to 0.9.7.0~dev
as suggested by @msprotz, any 0.9.6.x version will not
be tagged off branch master, but off branch v0.9.6+
which will not be merged with master, but will
consist of cherry-picks from master.

so, similarly, once 0.9.7.0 is released, we will create
a 0.9.7+ branch first, then change 0.9.7.0~dev to
0.9.7.0 in those files *there*, not in master.
Instead, we will bump to 0.9.8.0~dev in master.