tactics: mark all primitives with strong_reduction_ok=true
This should be safe, even for primitives such as `fresh : unit -> Tac
int` since they must be reified an applied to a proofstate in order
to run. In that case, if we're strongly reducing, the proofstate will
likely be a variable and will fail to be unembedded, and the primop will
not reduce.