tc: rel: tracking implicits that were being lost
Calling solve_t might return new implicits, which we must again keep
track of. To make sure we don't duplicate the implicits we're tracking,
we do the call with an empty list of implicits, and then we append the
result to the implicits we were already carrying.

c.f. #1443 and #1523