StackInline now implies inline_for_extraction noextract
It is actually fairly difficult to syntactically detect what constitutes
a StackInline function or not, due to higher order, type-level
normalization, etc. Furthermore, StackInline functions should be
normalized by F*, not by KreMLin.

As such, this change now warns (to be made fatal) about StackInline
functions that are not marked as inline_for_extraction noextract. The
motivation behind this is that some functions that ought to be flagged
as StackInline currently are not, leading to some C code that shouldn't
be generated showing up and triggering C compiler warnings.