normalizer: cfg: be more efficient in computing psteps set
When the extra_steps/plugins for the normalizer are non-empty, it can
take a whole bunch of time to traverse these lists and turn them into
psmap's, and this needs to be done on every. call to `config'` (which
are a LOT!).

So, let's memoize them. These lists seldom ever change, so this saves a
bunch of time.