[LSP] publishDiagnostics after didOpen
Take the opportunity to rewire the types to be simpler. On didOpen, if
for instance, some identifier is undeclared, it will appear with a
swiggly red underline. Until this error is fixed, all of Lsp goes down,
and no features are available. The next step is to invoke full_lax on
didSave.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>