Cleanup / small improvement over regionals.
Change the definition of the buffer regional to actually use the
parameter instead of relying on closures which most likely did not
compile to C in the first place. Also provides an example for how to use
the double-parameter style to specialize functions and therefore skip
the intermediary let-binding marked as inline_for_extraction.