tosyntax: fixing #829 and #1903 properly
The main problem is that top-level patterns that did not bind any
variable were not being checked for completeness. That is,

  let None = Some 42

was being desugared roughly to

  let uu___X = Some 42

and nothing else, and hence accepted. Instead, if there were any
variables bound by the patterns we would see a declaration to project
them that would fail to typecheck (due to an incomplete match).

So, instead of forbidding using constants on top-level patterns, or
whatever, let's just always generate a "projection" declaration which
will check for completeness of the match. In this case, it would be

  let uu___X = match Some 42 with | None -> ()

which would be properly rejected.

Fixes #829
Fixes #1228
Fixes #1903