tosyntax: restore check that type ascriptions on patterns are only over variables
This was removed when we made desugaring not set the types of variables
from the ascriptions (c2dd29eb5f6ed9def7f1c3c3755fbf24bb637cbe). It
needs be in place, however, since otherwise patterns such as this one

  match None with
  | (None : option int) ->  ()

explode with a variable not found error, instead of being promptly
rejected.

Note, however, that we need to allow things such as:

  let f ((x, y) : int & int) = E

Which are then desugared (correctly already) to something similar to

  let f (r : int & int) =
    let (x, y) = r in
    E

So we do *not* fail when
1) the ascription is at the top of the pattern, and
2) we are checking a binding pattern from a Let or Lam (but not a
   Match)

Fixes #1954