tosyntax/tc: make expect_failure work for desugaring errors too
At desugaring time, if an expect_failure definition fails to desugar,
we check that the errors it raises match those requested (if any), and
succeed if so. If the errors do not match, we fail.

However if the declaration *doesn't* fail, we do nothing. We kick it
forward to the typechecker who will check for typing errors.

This allows to have easy regression tests on desugaring features, while
not needing to introduce a new attribute for them nor disturbing the
behaviour of expect_failure for typechecking.

Note that declarations like this:

  let None = Some ()

will be expanded to TWO declarations

  let uu___0 = Some ()
  let uu___1 = match uu___0 with | None -> ()

The first of which will succeed (even typechecking), and where the
second will fail. So we desugar the first declaration into a *single*
Sig_group sigelt containing the two elaborated uu___0 and uu___1. The
Sig_group is the one that will contain the expect_failure attribute, so
the typechecker will check that the the group itself fails to typecheck,
as expected.

So, morally, we expand

  [@expect_failure]
  let None = Some ()

To something like the following (if we had surface syntax for Sig_group)

  [@expect_failure]
  Sig_group [
    let uu___0 = Some ()
    let uu___1 = match uu___0 with | None -> ()
  ]