normalizer: fix reification of bind for non-total effects
We were reifying

  reify (let x = e in f x)

as

  bind (reify e) (fun x -> reify (f x))

But this is not correct since `reify e` will be effectful
when the effect is non-Tot, and should be let-bound instead,
as in:

  let h : EFF _ = reify e in
  bind h (fun x -> reify (f x))

Where EFF is either Tot or Div, according to whether the effect
is total or not.

This causes tactics to get stuck since its effectful lets never reduce,
to be addresed in subsequent commits.