fstar: universal: do not update timestamp if extracting
I missed this. If we build A.fst.checked and B.fst.checked and then
extract A, we will invalidate B.fst.checked's timestamp causing a
rebuild (which won't do much except updating the timestamp, but takes
time). So don't update if we're extracting (the ml files are
freshly written every time anyway.)