syntax: subst: fix possible leak of Tm_delayed on compress
It was possible for compress to return a Tm_delayed node when it entered
the Tm_name case and the substitution on the node mapped the variable to
a Tm_delayed node.

The error comes since, at a first glance, `subst'` appears to never
return a Tm_delayed when called on `Tm_type`, `Tm_bvar` and `Tm_name`,
which are the cases on which its called in `push_subst`. However this
assumption is broken when the substitution itself contains a Tm_delayed.

So just recurse in compress. Hopefully the performance impact is
negligible, and this makes it evident that we never return a Tm_delayed.
Also, leave a commented-out check for this property for the future.