tc: norm: elim_delayed_subst: descend into lbnames
And bv sorts, and universes, and take care for memoized free variables.

The problem was posted in #2086, and I minimized it to something like:

  let b_hd = sub b 0ul 1ul in
  let b_tl = offset b 1ul in
  admit ()

What was happening was that a meta implicit for sub was being created
and substituted into the continuation (correctly) and hence into the
sort of `b_tl`, also correctly. This meta implicit is later solved to
`false`, and `elim_delayed_subst` is meant to replace the uvar node
with the solution so we can ditch the UF graph and write the term to
a file, but it was not descending into the sort of `b_tl`, hence the
node stayed. The node contains an `env` in the uvar, and it cannot be
serialized since it contains closures, hence the blowup.

I think only the lbname part will fix the issue, but the rest is going
to blow up in our face eventually. I also expanded some matches with
default cases into fully spelling out the cases (e.g. cflags) so when
something changes we are forced to revisit the logic instead of silently
defaulting.

This was not fun at all.

Fixes #2086