FStar.List.Tot.Base: improve some specs, remove eqtype restriction
We can use `memP` in `filter`, since the membership is used only for
specificational purposes. Hence there's no need to restrict the spec to
only work on eqtypes.

This would break some proofs that work only over `mem`, since
`memP` is a priori unrelated to it. There's a lemma `mem_memP` in
`FStar.List.Tot.Properties` stating their equivalence, but without a
pattern. This commit adds a pattern.