tc: fix lookup of recursive lids
We were instantiating the tscheme, but we must instead return
exactly the same universe names to be consistent. There is no
universe-polymorhic recursion. Also fix TcInductive to open the scheme
with the universe names before pushing into the environment.

I've unfolded tscheme to (univ_names * typ) in the syntax and added
a comment about it.

Fixes (part of?) #2106