tactics/reflection: use interfaces in ulib/ for the internal modules that are exposed
This is a lot cleaner, and makes extraction easier as one does
not have to explicitly exclude these. See the simplifications in
ulib/ml/Makefile.include.

Sadly, doing this will make the bootstrapping process pick them up
as interfaces of the internal modules, since they share a name. For
instance, taking ulib/FStar.Reflection.Data.fsti to be the interface of
src/reflection/FStar.Reflection.Data.fs, which is not intended. This
won't fly, since many more functions are exposed internally than to
userspace.

As a bandaid, we write an internal interface for these modules too.
We just copy the implementation file to it, we don't need anything
fancy. The duplication is annoying, but can be improved later, and much
better than the module mess for plugins.