ulib: some changes to Squash and IndefiniteDescription
Motivated by a discussion on Slack where we noticed we could not define
a function from [a -> GTot b] into [GTot (a -> b)], which is the axiom
of choice for GTot functions.

This commit tweaks the type of [push_squash] to return a squashed total
function (instead of a ghost function. The [push_squash] axiom is very
similar to what we need here, so we can repurpose it via indefinite
description to obtain the [GTot] choice.

Further, this commit simplifies the indefinite description axiom into
simply a function of type [squash a -> GTot a]. This is already provable
with the current formulation (by introducing a dummy existential and
obtaining its witness), but it seems more natural to state it like this.
The old versions are all proven with this axiom, some of them in simpler
ways.