syntax: dsenv: reimplement shorten_lid by trimming of prefixes
shorten_lid worked by taking the path of a lid, shortening it as per the
current modules in scope, and reattaching the shortened path to the lid.
However this causes issues since it is not taking shadowing caused by
definitions (see #2702).

This patch reimplements shorten_lid by leveraging lid resolution. To
shorten a name such as A.B.C.x, we try to resolve (in order) x, C.x,
B.C.x and A.B.C.x, and use the first one that works.

Also considers module abbreviations, if there is a `module T = A.B` in
scope, then any lid under A.B, for example A.B.C.D.foo *will* use the
abbreviation, but attempt to shorten the suffix, so the candidates are,
in order, `A.B.foo`, `A.B.C.foo`, and `A.B.C.D.foo`. This also uses the
most specific abbreviation available.

Fixes #2702