ulib: reflection: update after exposing ascriptions
This required a few tweaks.

1- Introduce `inspect_unascribe` and `inspect_ln_unascribe` and use them
in some existing functions, such as `term_as_formula`, to retain old
behavior. Idem for Steel.Effect.Common

2- Change `collect_app` to unascribe while checking. This is
probably what clients expect, and needed for `all_guards_solved` in
Steel.Effect.Common since it would otherwise loop on reaching an
ascription.

3- Due to that, `mk_app` and `collect_app` are no longer inverses, so
the lemma `mk_app_collect_inv` is gone.

4- `inspect_ln_unascribe` requires a certain transitivity of `smaller`
and `<<`, so introduce a lemma for it (`smaller_trans_r`).

5- Note, an analogous `smaller_trans_l` lemma would be the following:

    let smaller_trans_l
        (t1 : term) (tv2 : term_view) (t2 : term)
      : Lemma (requires (t1 << tv2 /\ tv2 `smaller` t2))
              (ensures t1 << t2)

which intuitively makes sense, but it cannot be proved (consider when
tv2 is `Tv_Const _`: the `smaller` hypothesis is trivial, so there is no
information about `t2`).