syntax: subst: remove bv sorts on deep_compress
The sorts on Tm_name and Tm_bvar nodes are unneded, and should be
removed at some point.

Currently, they may contain other terms and be affected by typechecking,
and would need to be compressed before writing them to a file (as issue
 #2684 shows)

As a bandaid, simply delete them on deep_compress: they are not needed
at all in the checked file, and we save the trouble of descending and
deep_compress'ing the sort, which also currently may fail.