reflection: reimplement term_eq to not leak term representation details
After discussion on PR #2645, this implements a variant of
`Syntax.Util.term_eq` in `Reflection.Basic`, which only looks
at information that is exposed via the view (inspect_ln). By
constructing it this way, we should not get any inconsistency wrt the
pack_inspect_inv lemma.

Note that this function now becomes just an optimization, since it is
implementable in userspace. This means it can _become a plugin_, but I'm
leaving that for a future commit.

At the same time, we expose the old `term_eq` via the Tac effect, with
the name `term_eq'`, so 1) migration is easier and 2) provide more power
to Tac programs, which should be able to inspect all details safely.