Makefiles: introduce a monitoring switch
With theses changes, passing MON=1 to the make invocation will prefix
all `fstar.exe` invocations with a `runlim` wrapper to save resource
consumption information into a `.runlim` file. One of these files is
generated for each makefile _rule_.

runlim has to be installed for this to work. And it *must* be a recent
version that supports the `-p` option. Without -p, a failure in an F*
invocation will not be noticed by make, and we end up with a false
success.

I suggest building and installing from https://github.com/arminbiere/runlim

The information can then be collected with the .scripts/res_summary.sh
script to produce a list of top offenders.