syntax/reflection: fix eq_tm to not break the view abstraction
Simply never return a meaningful result for quoted terms. We do not seem
to rely on it anywhere, and writing this function correctly is tricky.
Currently we cannot reuse any of eq_tm, Syntax.Util.term_eq nor
Reflection.Basic.term_eq as they ignore some details (such as name
of binders) that are exposed in the view, and hence this would
return a false positive which quickly gives rise to an inconsistency.

Cf. #2806.