Makefile.boot: take tests out of the library
Tests files (src/tests) make heavy use of top-level effects to
initialize a typechecker environment and load some definitions, among
other things.

If they are linked with everything else into fstar-lib, then
 1- the compiler will include them, which is a bit of bloat
 2- the compiler will be side effected by them causing some error
    message tests to fail as Tahina noticed.
 3- applications (which currently use this library) will also
    contain the tests and initialize them on program load, which
    is terrible.

This patch then moves all of the generation of FStar_Test_* modules
into ocaml/fstar-test instead of ocaml/fstar-lib. Dune takes care of the
rest. It would be nice for this way of splitting modules to be
standardized and somehow more easily configurable.

The actual patch is twofold: a SED edit of the .depend file, and a
change to the extraction rule to decide which directory a file goes to.
It does not really make sense to have to edit both... but there are
several limitations to our current Makefiles (see also #1978).