checkedfiles: suppress not found warning for --lax
If we call
    fstar.exe --cache_checked_modules --lax A.fst
we now get a warning that A.fst.checked.lax does not exist. This
does not really make sense as we are trying to build the file, but
`should_verify` return false, as we are in lax mode. So, use the weaker
`should_check` instead.

This removes the flurry of warning we get when lax-building the
libraries (though that process is about to disappear).