tactics: check for internal qualifiers when splicing
We cannot allow tactics to just set any qualifier. While some of them
will cause the typechecker to fail, in others it will just silently
accept them since it trusts the parser to not generate odd combinations,
such as `Effect` on a top-level value definition.

So, just raise an error if any internal qualifier appears in the
generated sigelt.