fstar: gzip checked files
Add output_zipped_value as an analogue to output_value, which calls gzip
on the resulting marshaled string before writing to the channel. Idem
for input_value.

This is not ideal, it would be better to just have a notion of a
compressed channel, but I couldn't find good support to do so in OCaml.
We could revisit it later, and perhaps use it to compress hint files
too.