introduce an `admit_termination` attribute
This attribute can be attached to a recursive group in order to disable
termination checking for it. It works over the entire group, propagating
the `admit_termination` to each sigelt. The typechecker, however,
can (and does) make the decision for each particular binding, so if
per-binding attribute syntax appears this can easily be used to admit
the termination of one particular definition in the group.

Use of this attribute triggers a Warn_WarnOnUse warning, which is silent
by default but picked up by --report_assumes.