ci: performance monitoring
This enables performance monitoring (as in building with
`RESOURCEMONITOR=1`) for CI builds.

It is conditioned by the `FSTAR_CI_RESOURCEMONITOR` variable is set
to `1` in the variables for the repository (Settings -> Secrets and
variables -> Actions -> Variables). It is currently already enabled even
without this PR.

In detail:
0) The workflow will pass the RESOURCEMONITOR=1 file to the Docker
   build command, this will cause our Makefiles to create .runlim
   files everywhere.
1) build-standalone.sh will, after running CI, create an `rmon/`
   directory and place relevant information there, including the summary
   generated by `.scripts/res_summary.sh` as well as all of the
   individual `.runlim` files ordered by path.
2) The workflow picks this directory up, and uploads two Github build
   artifacts, one with the summary and one with the tarball of all .runlim
   files.
3) Two lines added to the Slack message with total CPU and memory usage
   (to be seen how robust this measurement will be)

Github makes a .zip archive of an artifact when one tries to download
it (even if it contains a single file), which is annoying, but I don't
think there's an alternative currently.