Compiler: 'Stubbing' everything
To prevent circularities when using typeclasses within the compiler
(since we need to refer to tactic modules potentially from anywhere)
we stub all of the exposed interfaces in ulib, so the compiler does
not 'realize' they are the same modules. Extraction then takes care of
removing the 'Stubs' namespace component, essentially closing the loop.