extraction: extract the FStar.Stubs.* namespace to FStar.*
This allows to untie some dependency knots in the compiler, mostly
in support of using typeclasses/tactics in src/. This way, the
users of tactics in src/ (say defining a to_string instance in
FStar.Syntax.Print) could only depend on FStar.Stubs.Tactics.Builtins,
and not on the *actual* builtins with the implementations, which would
be a cycle.

This works as long as the dependency is only for tactic execution, as
the extracted code will not have such dependency.