ToSyntax: improve some ranges
With this change, the def-range for an inductive type or data
constructor points to the source location where the name appears in the
definition, and not the full recursive block. This means that hitting
F12 in the VSCode extension over a type or constructor will be more
precise.

cc @nikswamy