Enable memtrace on F*
This commit brings memtrace into F* to do memory profiling. It is
(always) linked to F*, but does not do anything unless the MEMTRACE
environment variable is set, so this should be zero-overhead (except
for a few kb of binary size) when memtrace is not used. When used, the
overhead is still supposed to be in the low digits of %.

With this enabled, we can just run

  MEMTRACE=trace.ctf fstar.exe ..

to obtain a trace.ctf file with information about the allocations.
This can then be inspected with the memtrace-viewer tool (installed
separately, since it's quite big).

I think having this enabled by default would be useful, and the cost
seems pretty low.

Some more info:
	https://blog.janestreet.com/finding-memory-leaks-with-memtrace/
	http://web.archive.org/web/20230528123929/https://blog.janestreet.com/finding-memory-leaks-with-memtrace/