deep_compress: remove some spurious warnings about Tm_names
deep_compress warned on reaching a Tm_name (when allow_uvars was false)
since they are not meant to be saved into checked files (as they are a
leak). However we are now using deep_compress more pervasively to check
that terms do not contain uvars, but in situations where names are
totally OK, such as the reflection API.

So, add a new toggle for deep_compress and make it not warn on Tm_names
in these cases.