feat: Add 'bv_div_unsafe' to support 'bv / bv' expressions.
This is useful when encoding LLVM IR semantics into FStar,
where we are only concerned about partial correctness in the cases
where the denominator is nonzero. Thus, we introduce a new
'bv_div_unsafe' operations, with both operands as bitvectors,
along with a soundness axiom which asserts that 'bv_div_unsafe'
behaves as 'bv_div' when the denominator is nonzero.

We add a 'bv_div_unsafe' lowering in the smt2 encoding
that mimics 'bv_div'.