Tc: fix sigelt pushing to be unique
We were routinely clashing names in the sigtab and overriding old
definitions. This patch makes it so that we never do that (raising an
error if it ever happens, say via a splice), except for the expected
case of overriding a `val` with a more specific definition (a `let`, or
a type/datacon).